requirements.txt
google-cloud-storage
xmltodict
